refactor: config 파일 패키지화
