+++ backend/src/main/java/develup/support/config/CorsConfig.java
@@ -1,4 +1,4 @@
-package develup.support;
+package develup.support.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
