+++ backend/src/main/java/develup/support/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package develup.support;
+package develup.support.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
